Rename permutation.rs to lookup_trace.rs
The module computes multiplicity traces for lookup arguments,
not permutations. The new name pairs with lookup.rs for discoverability.
