+++ backends/plonky3/src/lookup_trace.rs
